Huge improvement in RAG performance
We write in plain language descriptions of the source files and the user
query. The embedding models do not for obvious reasons know how to
represent whatever random variable/class/method names we have and
probably struggle with programming language/library specifics too.
But they work well with plain language.

Signed-off-by: Lauri Leukkunen <lauri.leukkunen@gmail.com>